Add separated KorUiSample example
