+++ samples/ui/build.gradle.kts
@@ -1,4 +1,6 @@
-import com.soywiz.korge.gradle.*
+import com.soywiz.korge.gradle.KorgeGradlePlugin
+import com.soywiz.korge.gradle.Orientation.LANDSCAPE
+import com.soywiz.korge.gradle.korge
 
 apply<KorgeGradlePlugin>()
 
@@ -6,8 +8,7 @@ korge {
 	id = "com.soywiz.sampleui"
 	name = "SampleUi"
 	description = "A sample using Korge and the gradle plugin"
-	orientation = com.soywiz.korge.gradle.Orientation.LANDSCAPE
-	jvmMainClassName = "SampleUiKt"
+	orientation = LANDSCAPE
 
 	targetDefault()
 }
