+++ samples/ui/src/commonMain/kotlin/main.kt
@@ -1,62 +1,28 @@
-import com.soywiz.klock.*
-import com.soywiz.korge.*
-import com.soywiz.korge.debug.*
-import com.soywiz.korge.html.*
-import com.soywiz.korge.input.*
-import com.soywiz.korge.service.process.*
-import com.soywiz.korge.tween.*
+import com.soywiz.klock.seconds
+import com.soywiz.korge.Korge
+import com.soywiz.korge.input.onClick
+import com.soywiz.korge.tween.get
+import com.soywiz.korge.tween.tween
 import com.soywiz.korge.ui.*
-import com.soywiz.korge.view.*
-import com.soywiz.korgw.*
-import com.soywiz.korim.bitmap.*
-import com.soywiz.korim.color.*
-import com.soywiz.korim.font.*
-import com.soywiz.korim.format.*
-import com.soywiz.korio.file.std.*
-import com.soywiz.korio.util.*
-import com.soywiz.korma.interpolation.*
+import com.soywiz.korge.view.position
+import com.soywiz.korgw.GameWindow
+import com.soywiz.korim.color.ColorTransform
+import com.soywiz.korim.color.transform
+import com.soywiz.korim.font.readBitmapFont
+import com.soywiz.korio.file.std.resourcesVfs
+import com.soywiz.korma.interpolation.Easing
 
-suspend fun main3() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI", bgcolor = Colors["#1c1e0e"]) {
-    val container = fixedSizeContainer(width, height, clip = true) { }
-    container.korui {
-        addChild(UiEditProperties(app, container, views))
-        /*
-        vertical {
-            horizontal {
-                preferredWidth = 100.percent
-                //minimumWidth = 100.percent
-                button("HELLO", {
-                    //minimumWidth = 50.percent
-                    preferredWidth = 70.percent
-                    //preferredHeight = 32.pt
-                })
-                button("WORLD", {
-                    preferredWidth = 30.percent
-                    preferredHeight = 32.pt
-                })
-            }
-            button("DEMO").apply {
-                visible = false
-            }
-            button("TEST")
-            checkBox("CheckBox", checked = true)
-            comboBox("test", listOf("test", "demo"))
-        }
-        */
-    }
-}
 
-suspend fun main() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI") {
-	val nativeProcess = NativeProcess(views)
+suspend fun main() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI", width = 800, height = 400) {
 
-    uiSkin = UISkin {
-        val colorTransform = ColorTransform(0.7, 0.9, 1.0)
-        this.uiSkinBitmap = this.uiSkinBitmap.withColorTransform(colorTransform)
-        this.buttonBackColor = this.buttonBackColor.transform(colorTransform)
-        this.textFont = resourcesVfs["uifont.fnt"].readBitmapFont()
-    }
+	uiSkin = UISkin {
+		val colorTransform = ColorTransform(0.7, 0.9, 1.0)
+		this.uiSkinBitmap = this.uiSkinBitmap.withColorTransform(colorTransform)
+		this.buttonBackColor = this.buttonBackColor.transform(colorTransform)
+		this.textFont = resourcesVfs["uifont.fnt"].readBitmapFont()
+	}
 
-    uiButton(256.0, 32.0) {
+	uiButton(256.0, 32.0) {
 		text = "Disabled Button"
 		position(128, 128)
 		onClick {
@@ -64,12 +30,11 @@ suspend fun main() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI
 		}
 		disable()
 	}
-    uiButton(256.0, 32.0) {
+	uiButton(256.0, 32.0) {
 		text = "Enabled Button"
 		position(128, 128 + 32)
 		onClick {
 			println("CLICKED!")
-			nativeProcess.close()
 		}
 		enable()
 	}
@@ -80,6 +45,7 @@ suspend fun main() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI
 			println(it.ratio)
 		}
 	}
+
 	uiScrollBar(32.0, 256.0, 0.0, 16.0, 64.0) {
 		position(64, 128)
 		onChange {
@@ -95,11 +61,15 @@ suspend fun main() = Korge(quality = GameWindow.Quality.PERFORMANCE, title = "UI
 		position(128, 128 + 64 + 32)
 	}
 
+	uiText("A simple Texbox") {
+		position(480, 32)
+	}
+
 	uiScrollableArea(config = {
 		position(480, 128)
 	}) {
 		for (n in 0 until 16) {
-            uiButton(text = "HELLO $n").position(0, n * 64)
+			uiButton(text = "HELLO $n").position(0, n * 64)
 		}
 	}
